+++ app/machine/instructions/halt.go
@@ -1,7 +1,7 @@
 package instructions
 
 import (
-	"github/com/lucasbn/sqlite-clone/app/btree"
+	"github/com/lucasbn/sqlite-clone/app/machine/common"
 	"github/com/lucasbn/sqlite-clone/app/machine/state"
 )
 
@@ -9,7 +9,7 @@ type Halt struct{}
 
 var _ Instruction = Halt{}
 
-func (Halt) Execute(s *state.MachineState, p btree.BTreeEngine) [][]int {
+func (Halt) Execute(s *state.MachineState, b common.BTreeEngine) [][]int {
 	s.Halted = true
 	return [][]int{}
 }
