+++ app/machine/instructions/instruction.go
@@ -1,10 +1,10 @@
 package instructions
 
 import (
-	"github/com/lucasbn/sqlite-clone/app/btree"
+	"github/com/lucasbn/sqlite-clone/app/machine/common"
 	"github/com/lucasbn/sqlite-clone/app/machine/state"
 )
 
 type Instruction interface {
-	Execute(s *state.MachineState, p btree.BTreeEngine) [][]int
+	Execute(s *state.MachineState, p common.BTreeEngine) [][]int
 }
