+++ app/machine/instructions/rewind.go
@@ -1,7 +1,7 @@
 package instructions
 
 import (
-	"github/com/lucasbn/sqlite-clone/app/btree"
+	"github/com/lucasbn/sqlite-clone/app/machine/common"
 	"github/com/lucasbn/sqlite-clone/app/machine/state"
 )
 
@@ -9,12 +9,12 @@ type Rewind struct {
 	Cursor uint64
 }
 
-var _ Instruction = Next{}
+var _ Instruction = Rewind{}
 
-func (rewind Rewind) Execute(s *state.MachineState, p btree.BTreeEngine) [][]int {
+func (rewind Rewind) Execute(s *state.MachineState, b common.BTreeEngine) [][]int {
 	s.CurrentAddress++
 
-	_, err := p.RewindCursor(rewind.Cursor)
+	_, err := b.RewindCursor(rewind.Cursor)
 	if err != nil {
 		panic(err)
 	}
